Added backtrace to SIGABRT because #178.
Free Software Free Society

To support commits by ctubio,
you can buy-me-a-drink with a small git tip at:
  1GitTipgxvKB3zjCLXRcSgDhC9pivkpc7u

I promise to drink chocolate milk in the development of the next commit.

To request new features or in case this commit breaks something for you,
please create a new github issue with all possible details,
but never share your API Keys!

+++ src/lib/mg.h
@@ -229,7 +229,7 @@ namespace K {
         UI::uiSend(uiTXT::FairValue, {{"price", mgFairValue}}, true);
       };
       static void cleanStdev() {
-        int periods = qpRepo["quotingStdevProtectionPeriods"].get<int>();
+        size_t periods = (size_t)qpRepo["quotingStdevProtectionPeriods"].get<int>();
         if (mgStatFV.size()>periods) mgStatFV.erase(mgStatFV.begin(), mgStatFV.end()-periods);
         if (mgStatBid.size()>periods) mgStatBid.erase(mgStatBid.begin(), mgStatBid.end()-periods);
         if (mgStatAsk.size()>periods) mgStatAsk.erase(mgStatAsk.begin(), mgStatAsk.end()-periods);
